Move firstrun components into a subdir for consistency
