+++ src/webextension/firstrun/components/wizard.js
